pyproject.toml
[tool.poetry]
name = "document_classifier"
version = "0.1.0"
description = "Document Classifier created from Kaggle Dataset"
authors = ["Mario Mesas <mmesasro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.1.1"
pandas = "^2.2.3"
numpy = "^2.2.3"
pydantic = "^2.10.6"
uvicorn = "^0.34.0"
fastapi = "^0.115.8"
opentelemetry-sdk = "^1.30.0"
opentelemetry-instrumentation-fastapi = "^0.51b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.30.0"
scikit-learn = "^1.6.1"
torch = "^2.6.0"
transformers = "^4.49.0"
requests = "^2.31.0"
streamlit = "^1.42.2"
pyyaml = "^6.0.2"
mlflow = "^2.20.3"


[tool.poetry.group.dev.dependencies]
lazypredict = "^0.2.13"
matplotlib = "^3.10.0"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
